Information Gathering using nmap
This code will gather the information using python module python-nmap. The code will take ip_address as input and response for the type of scan and will print the results on the screen. The code also contains option for banner grabbing for a specific port which will be taken as input.