A simple program for DOS attack
This program implements  the basic functionality of DOS attack, similar to Slowloris. 
With the help of Empty header, the code will try to engage the server for the time,
till incomplete requests are being made. 
Refer to Line 25 and 51 where we are sending only 1 line feed in place of 2 Line feeds, which marks the end of HTTP packet. Due to this the server will wait for other line feed for endless time.
The code will reinitiate the sockets which are being closed. 
Code written and executed By- Akhil Upadhyay PESU Cybersecurity group Head R&D dept. Head 
For any queries contact me on github- username akhil1010
The Author will not be responible for any type of malpractices or malicious activities anywhere using this code,as the code is written solely for the research and related purposes.